sửa logic thêm dịch vụ, sửa wifi

+++ src/main/java/dao/NhanVien_DAO.java
@@ -116,8 +116,8 @@ public class NhanVien_DAO {
 
     public boolean themNhanVien(NhanVien nv) {
         TaiKhoan_DAO tkDAO = new TaiKhoan_DAO();
-        String sql = "INSERT INTO NhanVien (maNhanVien, tenNhanVien, tenDangNhap, gioiTinh, ngaySinh, email, soDienThoai, trangThai, ngayBatDau, CCCD) "
-                + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?,?)";
+        String sql = "INSERT INTO NhanVien (maNhanVien, tenNhanVien, tenDangNhap, gioiTinh, ngaySinh, email, soDienThoai, trangThai, ngayBatDau, CCCD, diaChi) "
+                + "VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?,?, ?)";
 
         try (Connection con = ConnectDatabase.getConnection();
                 PreparedStatement stmt = con.prepareStatement(sql)) {
@@ -142,6 +142,7 @@ public class NhanVien_DAO {
             stmt.setString(8, "Đang làm việc");
             stmt.setDate(9, Date.valueOf(LocalDate.now()));
             stmt.setString(10, nv.getCCCD());
+            stmt.setString(11, nv.getDiaChi());
 
             int rows = stmt.executeUpdate();
             return rows > 0;
@@ -179,7 +180,7 @@ public class NhanVien_DAO {
         TaiKhoan_DAO tkDAO = new TaiKhoan_DAO();
         String vaiTroValue = nv.getTaiKhoan().getVaiTro().equalsIgnoreCase("Quản lý") ? "admin" : "employee";
         tkDAO.updateRole(nv.getTaiKhoan().getTenDangNhap(), vaiTroValue);
-        String sql = "UPDATE NhanVien SET tenNhanVien=?, gioiTinh=?, ngaySinh=?, email=?, soDienThoai=? WHERE CCCD=?";
+        String sql = "UPDATE NhanVien SET tenNhanVien=?, gioiTinh=?, ngaySinh=?, email=?, soDienThoai=?, diaChi=? WHERE CCCD=?";
         try (Connection conn = ConnectDatabase.getConnection();
                 PreparedStatement ps = conn.prepareStatement(sql)) {
             ps.setString(1, nv.getTenNhanVien());
@@ -187,7 +188,8 @@ public class NhanVien_DAO {
             ps.setDate(3, nv.getNgaySinh() != null ? Date.valueOf(nv.getNgaySinh()) : null);
             ps.setString(4, nv.getEmail());
             ps.setString(5, nv.getSoDienThoai());
-            ps.setString(6, nv.getCCCD());
+            ps.setString(6, nv.getDiaChi());
+            ps.setString(7, nv.getCCCD());
             return ps.executeUpdate() == 1;
         } catch (Exception ex) {
             ex.printStackTrace();
@@ -234,19 +236,20 @@ public class NhanVien_DAO {
 
     public boolean xoaNhanVienTheoCCCD(NhanVien nv) {
         TaiKhoan_DAO tkDAO = new TaiKhoan_DAO();
-        tkDAO.xoaTaiKhoanTheoTenDN(nv.getTaiKhoan().getTenDangNhap());
-        String sql = "DELETE FROM NhanVien WHERE CCCD = ?";
-        try (Connection conn = ConnectDatabase.getConnection();
-                PreparedStatement ps = conn.prepareStatement(sql)) {
-            ps.setString(1, nv.getCCCD());
-            boolean deleted = ps.executeUpdate() > 0;
-            if (deleted)
-                tkDAO.xoaTaiKhoanTheoTenDN(nv.getTaiKhoan().getTenDangNhap());
-            return deleted;
-        } catch (Exception e) {
-            // TODO: handle exception
+        boolean daXoaTaiKhoa = tkDAO.xoaTaiKhoanTheoTenDN(nv.getTaiKhoan().getTenDangNhap());
+        if (daXoaTaiKhoa) {
+            String sql = "DELETE FROM NhanVien WHERE CCCD = ?";
+            try (Connection conn = ConnectDatabase.getConnection();
+                    PreparedStatement ps = conn.prepareStatement(sql)) {
+                ps.setString(1, nv.getCCCD());
+                boolean deleted = ps.executeUpdate() > 0;
+                if (deleted)
+                    tkDAO.xoaTaiKhoanTheoTenDN(nv.getTaiKhoan().getTenDangNhap());
+                return deleted;
+            } catch (Exception e) {
+                // TODO: handle exception
+            }
         }
-        System.out.println("Xoa that bai o dbs nv");
         return false;
     }
